app/backend/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
pydantic==2.1.1
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
alembic==1.11.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.21.1
httpx==0.24.1
reportlab==4.0.4
Pillow==10.0.0
pandas==2.0.3
openpyxl==3.1.2